Added template navbar to cam.html and readded php headers back to other modules
Template navbar was added to cam.html.  Currently the dropdown doesn't work with the page at full size but does when the page is resized to be small horizontally. There seems to be a conflict with bootstrap as when i comment it out the dropdown starts working. Once dropdown is fixed the navbar will be stylized and not use filler buttons.